Add builder package_venv
This adds class PackageVenvBuilder which builds the virtualenv
in a temporary path, and then uses virtualenv-tools to update the
contents to match the provided install path.
This allows multiple concurrent Giftwrap jobs on a single machine
and does not require exclusive write access to an absolute path.
(eg. /opt/<company>/<project>)